+++ src/word_search_generator/word_search/word_search.py
@@ -130,7 +130,7 @@ class WordSearch(Game):
         )
 
         # setup words
-        word_set = WordSet()
+        word_set: WordSet = OrderedSet()
         if words:
             word_set.update(self._process_input(words))
         if secret_words:
@@ -153,12 +153,12 @@ class WordSearch(Game):
     @property
     def hidden_words(self) -> WordSet:
         """Words of type "hidden"."""
-        return WordSet(word for word in self.words if not word.secret)
+        return OrderedSet(word for word in self.words if not word.secret)
 
     @property
     def placed_hidden_words(self) -> WordSet:
         """Words of type "hidden" currently placed in the puzzle."""
-        return WordSet(word for word in self.placed_words if not word.secret)
+        return OrderedSet(word for word in self.placed_words if not word.secret)
 
     @property
     def unplaced_hidden_words(self) -> WordSet:
@@ -168,12 +168,12 @@ class WordSearch(Game):
     @property
     def secret_words(self) -> WordSet:
         """Words of type "secret"."""
-        return WordSet(word for word in self.words if word.secret)
+        return OrderedSet(word for word in self.words if word.secret)
 
     @property
     def placed_secret_words(self) -> WordSet:
         """Words of type "secret" currently placed in the puzzle."""
-        return WordSet(word for word in self.placed_words if word.secret)
+        return OrderedSet(word for word in self.placed_words if word.secret)
 
     @property
     def unplaced_secret_words(self) -> WordSet: